Only run npm install /app when starting non-production apps
The npm install functionality ran on each launch.  The installed
packages are already fetched during the build phase, hence there
should be no reason to fetch the npm packages again unless we are
running in development mode.

The earlier implementations also did an npm install of the image's
packages.  These should be bundled with the image and should not be
updated accidentally.  Hence we choose to only run the npm install on
build of the template image, not on the build of the docker image
which uses this template.