Use `exec` in boot.sh and Node.js scripts
Using `exec` here means that the called command will replace the shell
process used to invoke the command. This is a simple way to allow
signals (such as a SIGTERM sent by `docker stop`) to propagate from the
main process to the actual application process that may want to listen
to these signals.