add a "test" mode
this mode works differently from dev in that it creates a merged package.json
file, so we have both the dependencies of the template and of the app available
in case of overlapping keys the values from the app are preferred over those
from the template. This is typically a requirement to successfully run tests.

Merging is done using jq's multiply function and probably has some limitations I'm
currently not aware of. So far it seems to be working.