Support /config to be used in production builds
The use of /config now also works in production builds.  We copy over
`/config` ignoring error output because an empty /config folder (which
is defacto all config folders at this point in time) yields an error.

I prefer to have that piece of code in there because I suspect future
additions will not remember that that would be something that needs to
be added and will assume it would just work.