Support for /config folder (#36)
* Support use of /config in development mounted in /app/config

This adds support for /config in development.  It's mounted as a folder
in /app/config and can use refer to other files with relative paths.

* Support /config to be used in production builds

The use of /config now also works in production builds.  We copy over
`/config` ignoring error output because an empty /config folder (which
is defacto all config folders at this point in time) yields an error.

I prefer to have that piece of code in there because I suspect future
additions will not remember that that would be something that needs to
be added and will assume it would just work.