Allow config folder to overwrite in different format
The user's config folder is merged with the app's config folder.  A
constraint on overwrites was that the overwriting code had to be written
in the same extension as the code that was being overwritten.  We now
consider js, ts and coffee to solve the same need and thus allow you to
overwrite a .ts file with a .coffee file.