Switch from alpine to Debian bookworm for node inspector
Shifting to debian means we need to explicitly use bash.

When using the node debugger through chromium based browsers (such as
Brave) the node debugger fails to connect to the nodejs inspector.
Together with @erikap we inspected the packets and it seems nodejs does
not accept incoming connections on the ipv6 address as its hostname.  It
is unclear why it does so.  It does accept connections on the specific
ipv4 address 127.0.0.1.  Searching for solutions it seems Debian
Bookworm uses different basic settings (though it is unclear which those
are exactly) that do make this case work.

In case of alpine there seems to be an ipv6 incompatibility.  Explicitly
adding 127.0.0.1:9229 as an inspect target makes things work.
Explicitly binding an ipv4 port in Docker (making it drop the IPv6 port)
makes things work too.  Using an older NodeJS version (up to alpine
node:15) also makes it work.  We therefore conclude this is a
compatibility issue and decide not to dig further.

We have dug into the packages.  We have searched online.  People seem to
accept this is the state of NodeJS and lacking deep dive guides
regarding nodejs, this seems to be the most straight path forward.