Introduce beforeExit and switch to watchman
beforeExit allows to run a series of commands before exit occurs.  The function
is async.

- setExitHandler is an unpublished escape hatch not intended for public use
- exitHandler is not intended for public use but is necessary in start-server.js
  which has been given its own file

nodemon has been removed

watchman has its own set of issues.  It is not clear when they appear exactly
but the issues suggest our current approach may work in our current setting.
Because we have not seen any extra errors during testing we'll accept this as
being as good as it gets.  We switched because nodemon has some open issues and
we could not get it to reliably forward signals necessary to clean up state
during development.