WIP Mostly works with many TODO: to signal code to remove
This version largely works.

The current main issue seems to be when rapidly changing a source file.  It
seems we then land in a loop in which the build is never fully finished or the
server never fully starts.

Comments with TODO: should be removed.  They're just there so we can enable them
again during current debugging.